text-editor: move paste handler to hook
