+++ src/serlo-editor/plugins/text/components/text-editor.tsx
@@ -1,5 +1,5 @@
 import clsx from 'clsx'
-import React, { useMemo, useEffect, useCallback } from 'react'
+import React, { useMemo, useEffect } from 'react'
 import { createEditor, Node, Transforms } from 'slate'
 import { Editable, ReactEditor, Slate, withReact } from 'slate-react'
 
@@ -8,29 +8,16 @@ import { Suggestions } from './suggestions'
 import { TextLeafRenderer } from './text-leaf-renderer'
 import { TextToolbar } from './text-toolbar'
 import { useEditableKeydownHandler } from '../hooks/use-editable-key-down-handler'
+import { useEditablePasteHandler } from '../hooks/use-editable-paste-handler'
 import { useEditorChange } from '../hooks/use-editor-change'
 import { useRenderElement } from '../hooks/use-render-element'
 import { useSuggestions } from '../hooks/use-suggestions'
 import { useTextConfig } from '../hooks/use-text-config'
 import type { TextEditorConfig, TextEditorState } from '../types/config'
-import { sliceNodesAfterSelection } from '../utils/document'
 import { useEditorStrings } from '@/contexts/logged-in-data-context'
-import { showToastNotice } from '@/helper/show-toast-notice'
 import { HoverOverlay } from '@/serlo-editor/editor-ui'
 import { useFormattingOptions } from '@/serlo-editor/editor-ui/plugin-toolbar/text-controls/hooks/use-formatting-options'
-import { isSelectionWithinList } from '@/serlo-editor/editor-ui/plugin-toolbar/text-controls/utils/list'
 import type { EditorPluginProps } from '@/serlo-editor/plugin'
-import { editorPlugins } from '@/serlo-editor/plugin/helpers/editor-plugins'
-import {
-  selectDocument,
-  selectParent,
-  insertPluginChildAfter,
-  selectMayManipulateSiblings,
-  runReplaceDocumentSaga,
-  useAppDispatch,
-  store,
-} from '@/serlo-editor/store'
-import { EditorPluginType } from '@/serlo-editor-integration/types/editor-plugin-type'
 
 export type TextEditorProps = EditorPluginProps<
   TextEditorState,
@@ -40,8 +27,6 @@ export type TextEditorProps = EditorPluginProps<
 // Regular text editor - used as a standalone plugin
 export function TextEditor(props: TextEditorProps) {
   const { state, id, editable, focused, containerRef } = props
-
-  const dispatch = useAppDispatch()
 
   const textStrings = useEditorStrings().plugins.text
 
@@ -69,6 +54,10 @@ export function TextEditor(props: TextEditorProps) {
     showSuggestions,
     previousSelection,
     state,
+  })
+  const handleEditablePaste = useEditablePasteHandler({
+    editor,
+    id,
   })
 
   // Workaround for setting selection when adding a new editor:
@@ -121,119 +110,6 @@ export function TextEditor(props: TextEditorProps) {
       clearTimeout(timeout)
     }
   }, [editor, focused])
-
-  const handleEditablePaste = useCallback(
-    (event: React.ClipboardEvent) => {
-      const isListActive = isSelectionWithinList(editor)
-
-      const document = selectDocument(store.getState(), id)
-      if (!document) return
-
-      const mayManipulateSiblings = selectMayManipulateSiblings(
-        store.getState(),
-        id
-      )
-      if (!mayManipulateSiblings) return
-
-      const parentPluginType = document.plugin
-
-      const files = Array.from(event.clipboardData.files)
-      const text = event.clipboardData.getData('text')
-
-      // Handle pasted images or image URLs
-      if (files?.length > 0 || text) {
-        const imagePlugin = editorPlugins.getByType(EditorPluginType.Image)
-        if (!imagePlugin) return
-
-        const imagePluginState =
-          imagePlugin.onFiles?.(files) ?? imagePlugin.onText?.(text)
-
-        if (imagePluginState !== undefined) {
-          if (isListActive) {
-            showToastNotice(textStrings.noElementPasteInLists, 'warning')
-            return
-          }
-
-          insertPlugin(EditorPluginType.Image, imagePluginState)
-          return
-        }
-      }
-
-      if (text) {
-        // Handle pasted video URLs
-        const videoPluginState = editorPlugins
-          .getByType(EditorPluginType.Video)
-          ?.onText?.(text)
-        if (videoPluginState !== undefined) {
-          event.preventDefault()
-
-          if (isListActive) {
-            showToastNotice(textStrings.noElementPasteInLists, 'warning')
-            return
-          }
-
-          insertPlugin(EditorPluginType.Video, videoPluginState)
-          return
-        }
-
-        // Handle pasted geogebra URLs
-        const geogebraPluginState = editorPlugins
-          .getByType(EditorPluginType.Geogebra)
-          ?.onText?.(text)
-        if (geogebraPluginState !== undefined) {
-          event.preventDefault()
-
-          if (isListActive) {
-            showToastNotice(textStrings.noElementPasteInLists, 'warning')
-            return
-          }
-
-          insertPlugin(EditorPluginType.Geogebra, geogebraPluginState)
-          return
-        }
-      }
-
-      function insertPlugin(
-        pluginType: string,
-        { state }: { state?: unknown }
-      ) {
-        const isEditorEmpty = Node.string(editor) === ''
-
-        if (mayManipulateSiblings && isEditorEmpty) {
-          dispatch(runReplaceDocumentSaga({ id, pluginType, state }))
-          return
-        }
-
-        const parent = selectParent(store.getState(), id)
-        if (!parent) return
-
-        const slicedNodes = sliceNodesAfterSelection(editor)
-
-        setTimeout(() => {
-          if (slicedNodes) {
-            dispatch(
-              insertPluginChildAfter({
-                parent: parent.id,
-                sibling: id,
-                document: {
-                  plugin: parentPluginType,
-                  state: slicedNodes,
-                },
-              })
-            )
-          }
-          dispatch(
-            insertPluginChildAfter({
-              parent: parent.id,
-              sibling: id,
-              document: { plugin: pluginType, state },
-            })
-          )
-        })
-      }
-    },
-    [dispatch, editor, id, textStrings]
-  )
 
   return (
     <Slate
